Rename function with autoimport enabled
