--- rules/coding-style/tests/Rector/Use_/UseFunction/Source/functions.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Deployer;
-
-function set($name, $value)
-{
-
-}
