+++ rules/renaming/tests/Rector/FuncCall/RenameFunctionAutoimportRector/config/configured_rule.php
@@ -1,9 +1,7 @@
 <?php
 
-declare(strict_types=1);
-
-use Rector\CodingStyle\Tests\Rector\Use_\UseFunction\Source\UseFunctionRector;
 use Rector\Core\Configuration\Option;
+use Rector\Renaming\Rector\FuncCall\RenameFunctionRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
@@ -11,5 +9,10 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $parameters->set(Option::AUTO_IMPORT_NAMES, true);
 
     $services = $containerConfigurator->services();
-    $services->set(UseFunctionRector::class);
+    $services->set(RenameFunctionRector::class)
+        ->call('configure', [[
+            RenameFunctionRector::OLD_FUNCTION_TO_NEW_FUNCTION => [
+                'view' => 'Laravel\Templating\render',
+            ],
+        ]]);
 };
